Type annotations, formatting, minor changes for Julia 1.0

+++ src/length.jl
@@ -15,12 +15,18 @@ Out: `length` [m], Line length
 Ported from .m file provided by Jonathon Martin.
 Original code by Mads Almassalkhi.
 """
-function estimate_length(S_base, V_base, R_pu, R_cond, bundle)
-    R_base = V_base^2/S_base
-    R_ohms = R_pu*R_base
-    line_length = R_ohms/(R_cond/bundle)
+function estimate_length(
+    S_base::Float64,
+    V_base::Float64,
+    R_pu::Float64,
+    R_cond::Float64,
+    bundle::Int64
+    )
+    R_base = V_base^2 / S_base
+    R_ohms = R_pu * R_base
+    line_length = R_ohms / (R_cond / bundle)
     # Ensure length estimate is positive
-    default = 1e-4*1609.3
+    default = 1e-4 * 1609.3
     line_length <= 0 && (line_length = default)
     return line_length
 end
